Replace phantomjs with chrome
Phantomjs is based on a tired, six year old WebKit engine which
doesn't support of standards we rely on (e.g. Flexbox). While we use
it to try to test our feature specs in a real-world way, it isn't
seeing a realistic visual of our app which is causing us to write
tests differently around it.

While there is a beta version of phantomjs that upgrades its engine
_slightly_, it is not present as a binary and the project itself is
no longer supported by the author.

This commit swaps out phantomjs for Chrome, which can now be run in
'headless' mode for testing via Selenium. There are some caveats here:

- The current Google-released and Debian-hosted versions of
  Chromedriver are both 2.30 which has a bug preventing the use of
  keycodes, which Capybara relies on for clicking. It has now been
  fixed upstream, and will be present in 2.31 which does not have a
  release date yet. Once it is available, we should stop relying on
  this GitHub-hosted contrib binary and use something more official,
  which will remove the need for unlinked depenancies such as
  libglib2.0.0
- We need a relatively modern version of Chrome, so we
  have to use Debian unstable/sid in addition to our oldstable/jessie
  distro. We use version pinning to ensure that apt get doesn't start
  fetching from unstable unless we specifically ask it too.