Install phantomJS via NPM rather than a binary
PhantomJS is old, rickety, and not well suited to our purposes (for 
instance, it doesn't support Flexbox, which we use). It's also no longer
being actively developed.

The alternatives, such as running Chrome headless, aren't up and running
yet. Until they are, we're going to continue using PhantomJS.

However, newest binaries of PhantomJS have a problem being fetched from
from their hosting on Bitbucket. It's not clear if this is a problem 
with us, the network, phantomJS or bitbucket, but many other people are
also reporting the issue with various degrees of resolution. The result
is that we can't build new docker images, because they can't install the
PhantomJS binary from source, which means we can't update our images 
with new software like ImageMagick.

This commit swaps the manual binary install with installing it via the
NPM wrapper. The `-g` flag is required to get it to `/usr/bin/phantomjs`
which is required for poltergeist and capybara to be able to find it.

This shouldn't actually change anything - the NPM wrapper just works out
what binary you need and gets it from the CDN - but it seems to be
making it work.