Upgrade to newest version of Node & Yarn
New Node comes with something called Corepack, which installs Yarn for
us rather than us having to get the installer ourselves and use it with apt-get.

We fix the yarn version to make sure it's always on the latest stable 
using `yarn set version` - this _might_ be optional? but the docs are
very unclear on that point and better safe than sorry.

This upgrade by necessity means upgrading NPM; new npm needs to be run 
_after_ setting the  WORKDIR so `npm install` is now further down the
list, but otherwise hasn't changed at all.